requirements-dev.txt
black==22.3.0
django_coverage_plugin~=1.6.0
factory-boy~=2.12.0
pytest~=5.0.1
pytest-cov~=2.7.1
pytest-django~=3.5.1

# Linting
pre-commit==2.9.2
